remove all database reliance
