+++ nodes/protocols/flow/network.go
@@ -2,7 +2,8 @@ package flow
 
 import (
 	"fmt"
-	"github.com/NubeDev/flow-eng/db"
+
+	"github.com/NubeDev/flow-eng/connections"
 	"github.com/NubeDev/flow-eng/node"
 	"github.com/NubeDev/flow-eng/schemas"
 	"github.com/NubeDev/flow-eng/services/mqttclient"
@@ -14,7 +15,7 @@ type Network struct {
 	*node.Spec
 	firstLoop                    bool
 	loopCount                    uint64
-	connection                   *db.Connection
+	connection                   *connections.Connection
 	mqttClient                   *mqttclient.Client
 	mqttConnected                bool
 	points                       []*point
@@ -50,10 +51,10 @@ func (inst *Network) setConnection() {
 		inst.setError(errMes, false, false)
 		return
 	}
-	var connection *db.Connection
+	var connection *connections.Connection
 	var connectionName = "flow framework integration over MQTT (dont not edit/delete)" // this name is set in rubix-edge-wires
 
-	connection, err = inst.GetDB().GetConnection(settings.Conn)
+	connection, err = inst.Connections().GetConnection(settings.Conn)
 	if err != nil {
 		errMes := fmt.Sprintf("flow-network error in getting connection: %+v. err:%s", settings.Conn, err.Error())
 		log.Errorf(errMes)
@@ -61,7 +62,7 @@ func (inst *Network) setConnection() {
 	}
 
 	if connection == nil {
-		connection, err = inst.GetDB().GetConnectionByName(connectionName)
+		connection, err = inst.Connections().GetConnectionByName(connectionName)
 		if err != nil {
 			errMes := fmt.Sprintf("flow-network error in getting connection name: %+v. err:%s", connectionName, err.Error())
 			log.Errorf(errMes)
