+++ nodes/protocols/flow/schema.go
@@ -3,6 +3,7 @@ package flow
 import (
 	"encoding/json"
 	"fmt"
+
 	"github.com/NubeDev/flow-eng/schemas"
 	"github.com/NubeIO/lib-schema/schema"
 	log "github.com/sirupsen/logrus"
@@ -17,28 +18,22 @@ const selectConnection = "Please add/select a MQTT connection"
 func (inst *Network) getConnectionsNames() (names []string, uuids []string) {
 	names = append(names, selectConnection)
 	uuids = append(uuids, selectConnection)
-	db := inst.GetDB()
-	if db != nil {
-		connections, err := inst.GetDB().GetConnections()
-		if err != nil {
-			log.Errorf("flow-networks get connections err %s", err.Error())
-			return nil, nil
-		}
-		for _, connection := range connections {
-			name := fmt.Sprintf("name:%s ip:%s port:%d", connection.Name, connection.Host, connection.Port)
-			names = append(names, name)
-			uuids = append(uuids, connection.UUID)
-		}
-		return names, uuids
+	connections, err := inst.Connections().GetConnections()
+	if err != nil {
+		log.Errorf("flow-networks get connections err %s", err.Error())
+		return nil, nil
 	}
-	log.Errorf("flow-networks failed to get db instance")
-	return nil, nil
-
+	for _, connection := range connections {
+		name := fmt.Sprintf("name:%s ip:%s port:%d", connection.Name, connection.Host, connection.Port)
+		names = append(names, name)
+		uuids = append(uuids, connection.UUID)
+	}
+	return names, uuids
 }
 
 func (inst *Network) buildSchema() *schemas.Schema {
-	names, uuids := inst.getConnectionsNames()
 	props := &nodeSchema{}
+	names, uuids := inst.getConnectionsNames()
 	props.Conn.Title = "connections"
 	if len(names) > 0 {
 		props.Conn.Default = names[0]
